% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_d}
\alias{c_d}
\title{Computes the probability density of a finitized distribution.}
\usage{
c_d(n, val, params, dtype)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{val}{The value of the variable for which the probability density function is computed.}

\item{params}{Other parameters of the distribution. They are provided in a list with named items, where the name
of an item is the name of the parameter.}

\item{dtype}{The type of the distribution: Poisson, Binomial, NegativeBinomial, Logaritmic.}
}
\value{
a\code{NumericVector} with the values of the density for each value provide in \code{val}.
}
\description{
Computes the probability density of a finitized distribution.
}
