% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_printDensity}
\alias{c_printDensity}
\title{Generates a string representation of the probability density function.}
\usage{
c_printDensity(n, val, params, dtype, latex = FALSE)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{val}{The value of the variable for which the probability density function is computed.}

\item{params}{Other parameters of the distribution. They are provided in a list with named items, where the name
of an item is the name of the parameter.}

\item{dtype}{The type of the distribution: Poisson, Binomial, NegativeBinomial, Logaritmic.}

\item{latex}{If true it returns a Latex formatted string representation of the pdf, otherwise it returns
the string representation of the pdf as an R expression.}
}
\value{
a string representation of the pdf.
}
\description{
Generates a string representation of the probability density function.
}
\keyword{internal}
