% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{dbinom}
\alias{dbinom}
\title{The density for the finitized Binomial distribution.}
\usage{
dbinom(n, p, N, val = NULL)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{p}{The parameter of the finitized Binomial distribution - the success probability for each trial.\eqn{p \in [0,1]}}

\item{N}{The number of trials.\code{ n = 0, 1, 2, ...}}

\item{val}{A vector with the values of the variable for which the probability density is computed. If NULL, a data frame containing
all possible values, i.e. \code{{0 .. n}}, and the corresponding probabilities is returned.}
}
\value{
a \code{data.frame} object with two columns: \code{val} containing the values passed in parameter \code{val} and \code{prob}
containing the corresponding densities.
}
\description{
\code{dbinom(n, p, N, val)} computes the finitized Binomial density for \code{val}.
}
\examples{
library(finitization)
dbinom(4, 0.5, 4)

}
