% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{dlog}
\alias{dlog}
\title{The density for the Logarithmic distribution.}
\usage{
dlog(n, theta, val = NULL)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{theta}{The parameter of the finitized Logarithmic distribution.}

\item{val}{A vector with the values of the variable for which the probability density is computed. If NULL, a data frame containing
all possible values, i.e. \code{{0 .. n}}, and the corresponding probabilities is returned.}
}
\value{
a \code{data.frame} object with two columns: \code{val} containing the values passed in parameter \code{val} and \code{prob}
containing the corresponding densities.
}
\description{
\code{dlog(n, theta, val)} computes the finitized Logarithmic density for \code{val}.
}
\examples{
library(finitization)
dlog(4, 0.5, c(0,1,3))

}
