% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{dnegbinom}
\alias{dnegbinom}
\title{The density for the finitized Negative Binomial distribution.}
\usage{
dnegbinom(n, q, k, val = NULL)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{q}{The parameter of the finitized Negative Binomial distribution - the success probability for each trial.\eqn{q \in [0,1]}}

\item{k}{The number of failures until the experiment is stopped,\code{k > 0}.}

\item{val}{A vector with the values of the variable for which the probability density is computed. If NULL, a data frame containing
all possible values, i.e. \code{{0 .. n}}, and the corresponding probabilities is returned.}
}
\value{
a \code{data.frame} object with two columns: \code{val} containing the values passed in parameter \code{val} and \code{prob}
containing the corresponding densities.
}
\description{
\code{dbinom(n, q, k, val)} computes the finitized Negative Binomial density for \code{val}.
}
\examples{
library(finitization)
dnegbinom(4, 0.12, 4)

}
