% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{dpois}
\alias{dpois}
\title{The density for the finitized Poisson distribution.}
\usage{
dpois(n, theta, val = NULL)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{theta}{The parameter of the finitized Poisson distribution.}

\item{val}{A vector with the values of the variable for which the probability density is computed. If NULL, a data frame containing
all possible values, i.e. \code{{0 .. n}}, and the corresponding probabilities is returned.}
}
\value{
a \code{data.frame} object with two columns: \code{val} containing the values passed in parameter \code{val} and \code{prob}
containing the corresponding densities.
}
\description{
\code{dpois(n, theta, val)} computes the finitized Poisson density for \code{val}.
}
\examples{
library(finitization)
dpois(4, 0.5, c(0,1,3))

}
