% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findSolutions}
\alias{findSolutions}
\title{Solves the equation \eqn{pdf(n-1) = 0} and returns the maximum feasible parameter space.}
\usage{
findSolutions(func)
}
\arguments{
\item{func}{a function. The actual parameter will be \code{pdf(n-1)}.}
}
\value{
a vector with two values: the lower and the upper limits of the maximum feasible parameter space.
}
\description{
It is used to actually solve the equation \eqn{pdf(n-1) = 0} needed
the find the maximum feasible parameter space for a finitized distribution. After finding the solutions, it builds the interval
which represents the maximum feasible parameter space.
}
