% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finitization.R
\docType{package}
\name{finitization}
\alias{finitization}
\title{A package that implements the finitization of probability distribution functions.}
\description{
This package contains functions that implement the concept of finitization of the probability distribution functions.
It implements four finitized distributions: Poisson, Logarithmic, Binomial and Negative Binomial.
}
\details{
finitization: A package that implements the finitization of probability distribution functions.
}
