% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{getBinomialMFPS}
\alias{getBinomialMFPS}
\title{Maximum feasible parameter space for the finitized Binomial distribution.}
\usage{
getBinomialMFPS(n, N)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{N}{The number of trials.}
}
\value{
A vector of two elements, where the first element is the lower limit and the second element is the upper limit of the maximum feasible parameter space.
}
\description{
\code{getBinomialMFPS(n, N)} computes and returns the maximum feasible parameter space for the finitized Binomial distribution.
}
\examples{
library(finitization)
getBinomialMFPS(2, 4)

}
