% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{getLogarithmicMFPS}
\alias{getLogarithmicMFPS}
\title{Maximum feasible parameter space for the finitized Logarithmic distribution.}
\usage{
getLogarithmicMFPS(n)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}
}
\value{
A vector with two elements where the first element is the lower limit and the second is the upper limit.
}
\description{
\code{getLogarithmicMFPS(n)} computes and returns the maximum feasible parameter space for the finitized Logarithmic distribution.
}
\examples{
library(finitization)
getLogarithmicMFPS(4)

}
