% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogarithmicType}
\alias{getLogarithmicType}
\title{The Logarithmic distribution type.}
\usage{
getLogarithmicType()
}
\value{
the Logarithmic distribution type.
}
\description{
The Logarithmic distribution type.
}
