% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{getNegativeBinomialMFPS}
\alias{getNegativeBinomialMFPS}
\title{Maximum feasible parameter space  for the finitized Negative Binomial distribution.}
\usage{
getNegativeBinomialMFPS(n, k)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{k}{The number of failures until the experiment is stopped,\code{k > 0}.}
}
\value{
A vector with two elements where the first element is the lower limit of the maximum feasible parameter space
and the second is the upper limit.
}
\description{
\code{getNegativeBinomialMFPS(n, k)} computes and returns the maximum feasible parameter space for the finitized Negative
Binomial distribution with parameter \code{k}.
}
\examples{
library(finitization)
getNegativeBinomialMFPS(2, 4)

}
