% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNegativeBinomialType}
\alias{getNegativeBinomialType}
\title{The Negative Binomial distribution type.}
\usage{
getNegativeBinomialType()
}
\value{
the Negative Binomial distribution type.
}
\description{
The Negative Binomial distribution type.
}
\keyword{internal}
