% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MFPS_pdf}
\alias{MFPS_pdf}
\title{Computes the expression of the \code{pdf(n-1)} needed to compute the maximu feasible parameter space.}
\usage{
MFPS_pdf(n, params, dtype)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{params}{Other parameters of the distribution. They are provided in a list with named items, where the name
of an item is thename of the parameter.}

\item{dtype}{The type of the distribution: Poisson, Binomial, NegativeBinomial, Logaritmic.}
}
\value{
the expression of the \code{pdf(n-1)}.
}
\description{
Computes the expression of the \code{pdf(n-1)} needed to compute the maximu feasible parameter space.
}
