% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{pbinom}
\alias{pbinom}
\title{The cumulative distribution function (CDF) for the finitized Binomial distribution.}
\usage{
pbinom(n, p, N, val = NULL, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{p}{The success probability for each trial (0 <= p <= 1).}

\item{N}{The number of trials.}

\item{val}{A vector of values at which the CDF is computed. If \code{NULL},
a data frame containing all possible values (from 0 to n) and the corresponding cumulative probabilities is returned.}

\item{lower.tail}{Logical; if TRUE (default) probabilities are computed as \eqn{P(X \le x)}; if FALSE, as \eqn{P(X > x)}.}

\item{log.p}{Logical; if TRUE, the cumulative probabilities are returned on the logarithmic scale.}
}
\value{
If \code{val} is provided, a numeric vector of cumulative probabilities (or their logarithms if \code{log.p = TRUE}) is returned.
Otherwise, a \code{data.frame} with columns \code{val} and \code{cdf} is returned.
}
\description{
\code{pbinom(n, p, N, val, lower.tail, log.p)} computes the CDF for the finitized Binomial distribution at the given value(s).
}
\examples{
library(finitization)
pbinom(4, 0.5, 4, val = c(0, 2, 4))
pbinom(4, 0.5, 4, lower.tail = FALSE)
pbinom(4, 0.5, 4, log.p = TRUE)

}
