% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{ppois}
\alias{ppois}
\title{The cumulative distribution function (CDF) for the finitized Poisson distribution.}
\usage{
ppois(n, theta, val = NULL, log.p = FALSE, lower.tail = TRUE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{theta}{The parameter of the finitized Poisson distribution.}

\item{val}{A vector of values at which the CDF is computed. If \code{NULL},
a data frame containing all possible values (0, 1, ..., n) and their cumulative probabilities is returned.}

\item{log.p}{Logical; if TRUE, the cumulative probabilities are returned on the logarithmic scale.}

\item{lower.tail}{Logical; if TRUE (default) probabilities are calculated as \eqn{P(X \le x)}; if FALSE, as \eqn{P(X > x)}.}
}
\value{
If \code{val} is provided, a numeric vector of cumulative probabilities (or their logarithms if \code{log.p = TRUE})
is returned. Otherwise, a \code{data.frame} with columns \code{val} and \code{cdf} is returned.
}
\description{
\code{ppois(n, theta, val, log.p, lower.tail)} computes the CDF for the finitized Poisson distribution at the given value(s).
}
\examples{
# For a finitized Poisson distribution with n = 4 and theta = 0.5:
ppois(n = 4, theta = 0.5, val = c(0, 2, 4))
ppois(n = 4, theta = 0.5, log.p = TRUE)
ppois(n = 4, theta = 0.5, val = c(0, 2, 4), lower.tail = FALSE)

}
