% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printDensity}
\alias{printDensity}
\title{Title
Prints the density function for finitized}
\usage{
printDensity(n, val, params, type, latex)
}
\arguments{
\item{latex}{}
}
\value{

}
\description{
Title
Prints the density function for finitized
}
\examples{

}
