% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{printFinitizedLogarithmicDensity}
\alias{printFinitizedLogarithmicDensity}
\title{The string representation of the probability density function for the finitized Logarithmic distribution.}
\usage{
printFinitizedLogarithmicDensity(n, val = NULL, latex = FALSE)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{val}{The value for which the pdf is printed. If \code{NULL}, the pdf for all values (0 ... n) is printed.}

\item{latex}{Logical; if TRUE the output is formatted in LaTeX.}
}
\value{
A character vector containing the string representation(s) of the pdf.
}
\description{
\code{printFinitizedLogarithmicDensity(n, val, latex)} computes and prints the string representation of the pdf.
}
\examples{
library(finitization)
printFinitizedLogarithmicDensity(4)

}
