% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{printFinitizedNegativeBinomialDensity}
\alias{printFinitizedNegativeBinomialDensity}
\title{The string representation of the probability density function for the finitized Negative Binomial distribution.}
\usage{
printFinitizedNegativeBinomialDensity(n, k, val = NULL, latex = FALSE)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{k}{The number of failures until the experiment is stopped,\code{k > 0}.}

\item{val}{The value of the variable for which the probability density function is printed. If NULL, this function computes the
string representation of the pdf for all possible values, i.e. \code{{0 .. n}}.}

\item{latex}{If TRUE, a string representation of the pdf formatted in Latex format is printed, otherwise this function prints
the string representation of the pdf as an R expression.}
}
\value{
This function silently returns a vector of type \code{\link[base]{character}} with the string representation of the pdf(s).
The length of the vector is the same with the length of the parameter \code{val}, i.e. one element for each value in \code{val}.
}
\description{
\code{printFinitizedNegativeBinomialDensity(n,k, val, latex)} computes and prints the string
representation of the probability density function for the finitized Negative Binomial distribution with parameter \code{k}.
}
\examples{
library(finitization)
printFinitizedNegativeBinomialDensity(4, 4)

}
