% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{qbinom}
\alias{qbinom}
\title{The quantile function for the finitized Binomial distribution.}
\usage{
qbinom(n, p, N, prob, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{p}{The success probability for each trial (0 <= p <= 1).}

\item{N}{The number of trials.}

\item{prob}{A vector of probabilities (or log-probabilities if \code{log.p = TRUE}) for which the quantiles are computed.
Each probability must be between 0 and 1.}

\item{lower.tail}{Logical; if TRUE (default) the input probabilities are interpreted as \eqn{P(X \le x)};
if FALSE, they are interpreted as \eqn{P(X > x)}.}

\item{log.p}{Logical; if TRUE, the probabilities in \code{prob} are assumed to be given on the logarithmic scale.}
}
\value{
A numeric vector of quantiles corresponding to the input probabilities. For each probability,
the quantile is defined as the smallest integer \(x\) (from 0 to n) such that the cumulative probability
is at least the provided value.
}
\description{
\code{qbinom(n, p, N, prob, lower.tail = TRUE, log.p = FALSE)} computes the quantile function for the finitized Binomial distribution.
}
\examples{
library(finitization)
qbinom(4, 0.5, 4, prob = c(0.1, 0.5, 0.9))
qbinom(4, 0.5, 4, prob = log(c(0.1, 0.5, 0.9)), lower.tail = TRUE, log.p = TRUE)
qbinom(4, 0.5, 4, prob = c(0.1, 0.5, 0.9), lower.tail = FALSE)

}
