% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{qlog}
\alias{qlog}
\title{The quantile function for the finitized Logarithmic distribution.}
\usage{
qlog(n, theta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{theta}{The parameter of the finitized Logarithmic distribution.}

\item{p}{A numeric vector of probabilities. For \code{lower.tail = TRUE} (default), these are interpreted as \eqn{P(X \le x)};
if \code{lower.tail = FALSE} they are interpreted as \eqn{P(X > x)}.}

\item{lower.tail}{Logical; if TRUE (default) the input probabilities are lower-tail probabilities, otherwise upper-tail probabilities.}

\item{log.p}{Logical; if TRUE, the probabilities in \code{p} are assumed to be given on the logarithmic scale.}
}
\value{
A numeric vector of quantiles corresponding to the input probabilities.
}
\description{
\code{qlog(n, theta, p, lower.tail, log.p)} computes the quantile function for the finitized Logarithmic distribution.
}
\examples{
library(finitization)
qlog(4, 0.1, p = c(0.1, 0.1, 0.9))
qlog(4, 0.1, p = c(0.1, 0.1, 0.9), lower.tail = FALSE)

}
