% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{qnegbinomial}
\alias{qnegbinomial}
\title{The quantile function for the finitized Negative Binomial distribution.}
\usage{
qnegbinomial(n, q, k, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{q}{The success probability for each trial (\eqn{q \in [0,1]}).}

\item{k}{The number of failures until the experiment is stopped, \code{k > 0}.}

\item{p}{A numeric vector of probabilities. For \code{lower.tail = TRUE} (default), these are interpreted
as \eqn{P(X \le x)}; if \code{lower.tail = FALSE}, they represent \eqn{P(X > x)}.}

\item{lower.tail}{Logical; if TRUE (default) the input probabilities are lower-tail probabilities,
otherwise they are upper-tail probabilities.}

\item{log.p}{Logical; if TRUE, the probabilities in \code{p} are assumed to be on the logarithmic scale.}
}
\value{
A numeric vector of quantiles corresponding to the input probabilities. Each quantile is the smallest integer
for which the cumulative probability is at least the given probability.
}
\description{
\code{qnegbinomial(n, q, k, p, lower.tail = TRUE, log.p = FALSE)} computes the quantile(s)
corresponding to the given probability(ies) for the finitized Negative Binomial distribution.
}
\examples{
library(finitization)
qnegbinomial(n = 4, q = 0.12, k = 4, p = c(0.1, 0.5, 0.9))
qnegbinomial(n = 4, q = 0.12, k = 4, p = c(0.1, 0.5, 0.9), lower.tail = FALSE)

}
