% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{qpois}
\alias{qpois}
\title{The quantile function for the finitized Poisson distribution.}
\usage{
qpois(n, theta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{theta}{The parameter of the finitized Poisson distribution.}

\item{p}{A numeric vector of probabilities. For \code{lower.tail = TRUE} (default), these are interpreted as \eqn{P(X \le x)};
if \code{lower.tail = FALSE} they are interpreted as \eqn{P(X > x)}.}

\item{lower.tail}{Logical; if TRUE (default) the input probabilities are lower-tail probabilities; otherwise, they are upper-tail.}

\item{log.p}{Logical; if TRUE the probabilities in \code{p} are assumed to be on the logarithmic scale.}
}
\value{
A numeric vector of quantiles corresponding to the input probabilities. Each quantile is defined as the
smallest integer \eqn{x} for which the cumulative probability is at least the given probability.
}
\description{
\code{qpois(n, theta, p, lower.tail, log.p)} computes the quantile(s) corresponding to the given probability(ies)
for the finitized Poisson distribution.
}
\examples{
# For a finitized Poisson distribution with n = 4 and theta = 0.5:
qpois(n = 4, theta = 0.5, p = c(0.1, 0.5, 0.9))
qpois(n = 4, theta = 0.5, p = c(0.1, 0.5, 0.9), lower.tail = FALSE)

}
