% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{rlog}
\alias{rlog}
\title{Random values generation for the finitized Logarithmic distribution.}
\usage{
rlog(n, theta, no)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 1.}

\item{theta}{The parameter of the Logarithmic distribution.}

\item{no}{The number of random values to be generated.}
}
\value{
A vector of integers containing random values generated from the finitized Logarithmic distribution.
}
\description{
\code{rlog(n, theta, no)} generates random values according to the finitized Logarithmic distribution.
}
\examples{
library(finitization)
rlog(2, 0.25, 10)

}
