% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rvalues}
\alias{rvalues}
\title{Random values generation.}
\usage{
rvalues(n, params, no, dtype)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{params}{Other parameters of the distribution. They are provided in a list with named items, where the name
of an item is the name of the parameter.}

\item{no}{The number of random values to be generated.}

\item{dtype}{The type of the distribution: Poisson, Binomial, NegativeBinomial, Logaritmic.}
}
\value{
a\code{NumericVector} with the random values generated.
}
\description{
Random values generation.
}
\keyword{internal}
