% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{dbinom}
\alias{dbinom}
\title{The density for the finitized Binomial distribution.}
\usage{
dbinom(n, p, N, val = NULL, log = FALSE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{p}{The success probability for each trial (must satisfy 0 <= p <= 1).}

\item{N}{The number of trials.}

\item{val}{A vector of values at which the density is computed. If \code{NULL},
a data frame containing all possible values (from 0 to n) and the corresponding densities is returned.}

\item{log}{Logical; if TRUE, the (natural) logarithm of the probabilities is returned.}
}
\value{
A \code{data.frame} with two columns: \code{val}, which contains the values, and \code{prob},
which contains the corresponding density (or log density if \code{log = TRUE}).
}
\description{
\code{dbinom(n, p, N, val, log)} computes the finitized Binomial density for each value in \code{val}.
}
\examples{
library(finitization)
dbinom(4, 0.5, 4)
dbinom(4, 0.5, 4, log = TRUE)

}
