% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{dlog}
\alias{dlog}
\title{The density for the Logarithmic distribution.}
\usage{
dlog(n, theta, val = NULL, log = FALSE)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{theta}{The parameter of the finitized Logarithmic distribution.}

\item{val}{A vector with the values of the variable for which the probability density is computed.
If \code{NULL}, a data frame containing all possible values, i.e. \code{0 ... n}, and the corresponding
probabilities is returned.}

\item{log}{Logical; if TRUE, the (natural) logarithm of the computed probabilities is returned.}
}
\value{
A \code{data.frame} object with two columns: \code{val} containing the values and
\code{prob} containing the corresponding densities (or their logarithms if \code{log = TRUE}).
}
\description{
\code{dlog(n, theta, val, log)} computes the finitized Logarithmic density for each value in \code{val}.
}
\examples{
library(finitization)
dlog(4, 0.1, c(0,1,3))
dlog(4, 0.1, c(0,1,3), log = TRUE)

}
