% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finitization.R
\docType{package}
\name{finitization}
\alias{finitization-package}
\alias{finitization}
\title{finitization: A package that implements the finitization of probability distribution functions.}
\description{
This package contains functions that implement the concept of finitization of probability distribution functions.
It includes finitized versions of the Poisson, Logarithmic, Binomial, and Negative Binomial distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bogdanoancea/finitization}
  \item Report bugs at \url{https://github.com/bogdanoancea/finitization/issues}
}

}
\author{
\strong{Maintainer}: Bogdan Oancea \email{bogdan.oancea@gmail.com} (\href{https://orcid.org/0000-0001-6987-5137}{ORCID})

}
\keyword{internal}
