% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{getBinomialMFPS}
\alias{getBinomialMFPS}
\title{Maximum feasible parameter space  for the finitized Binomial distribution.}
\usage{
getBinomialMFPS(n, N)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 0.}

\item{N}{The number of trials.\code{ n = 0, 1, 2, ...}}
}
\value{
A vector with two elements where the first element is the lower limit of the maximum feasible parameter space
and the second is the upper limit.
}
\description{
\code{getBinomialMFPS(n, N)} computes and returns the maximum feasible parameter space for the finitized Binomial distribution with parameter
\code{N}.
}
\examples{
library(finitization)
getBinomialMFPS(2, 4)

}
