% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom.R
\name{pnegbinom}
\alias{pnegbinom}
\title{The cumulative distribution function (CDF) for the finitized Negative Binomial distribution.}
\usage{
pnegbinom(n, q, k, val = NULL, log.p = FALSE, lower.tail = TRUE)
}
\arguments{
\item{n}{The finitization order. An integer > 0.}

\item{q}{The success probability for each trial (\eqn{q \in [0,1]}).}

\item{k}{The number of failures until the experiment is stopped, \code{k > 0}.}

\item{val}{A vector with the values at which the CDF is computed. If \code{NULL},
a data frame containing all possible values (0, 1, ..., n) and their cumulative probabilities is returned.}

\item{log.p}{Logical; if TRUE, the cumulative probabilities are returned on the logarithmic scale.}

\item{lower.tail}{Logical; if TRUE (default) probabilities are calculated as \eqn{P(X \le x)}; if FALSE, as \eqn{P(X > x)}.}
}
\value{
If \code{val} is provided, a numeric vector of cumulative probabilities (or their logarithms if \code{log.p = TRUE}) is returned.
Otherwise, a \code{data.frame} with columns \code{val} and \code{cdf} is returned.
}
\description{
\code{pnegbinom(n, q, k, val, log.p, lower.tail)} computes the CDF for the finitized Negative Binomial distribution at the given value(s).
}
\examples{
library(finitization)
pnegbinom(4, 0.12, 4, val = c(0, 2, 4))
pnegbinom(4, 0.12, 4, log.p = TRUE)
pnegbinom(4, 0.12, 4, val = c(0, 2, 4), lower.tail = FALSE)

}
