% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{rbinom}
\alias{rbinom}
\title{Random values generation for the finitized Binomial distribution.}
\usage{
rbinom(n, p, N, no)
}
\arguments{
\item{n}{The finitization order. An integer > 1.}

\item{p}{The success probability for each trial (0 <= p <= 1).}

\item{N}{The number of trials.}

\item{no}{The number of random values to be generated.}
}
\value{
An integer vector of length \code{no}, with random values drawn from the finitized Binomial distribution.
}
\description{
\code{rbinom(n, p, N, no)} generates random values according to the finitized Binomial distribution.
}
\examples{
library(finitization)
rbinom(2, 0.5, 2, 10)

}
