% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{rlog}
\alias{rlog}
\title{Random values generation  for the finitized Logarithmic distribution.}
\usage{
rlog(n, theta, no)
}
\arguments{
\item{n}{The finitization order. It should be an integer > 1.}

\item{theta}{The parameter of the Logarithmic distribution.}

\item{no}{The number of random values to be generated.}
}
\value{
\code{rlog} returns a vector of type \code{\link[base]{integer}} containing random values generated according to the finitized Logarithmic distribution.
The number of values is given by the parameter \code{no}.
}
\description{
\code{rlog(n, theta, no)} generates random values according to the finitized Logarithmic distribution with parameter \code{theta}.
}
\examples{
library(finitization)
rlog(2, 0.25, 10)

}
